Notice 수정
DTO 생성
서비스 및 레포지토리 수정

+++ src/main/java/com/fmi/nada/board/notice/NoticeRepository.java
@@ -2,7 +2,6 @@ package com.fmi.nada.board.notice;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 }
