+++ src/main/java/com/fmi/nada/board/notice/NoticeService.java
@@ -24,13 +24,13 @@ public class NoticeService {
     }
 
     //공지사항 글 상세 페이지에서 보여지는 공지사항 글
-    public Notice getNotice(Integer notice_idx) {
+    public Notice getNotice(Long notice_idx) {
         Optional<Notice> noticeOptional = noticeRepository.findById(notice_idx);
         return noticeOptional.get();
     }
 
     //공지사항 삭제
-    public void deleteNotice(Integer notice_idx) {
+    public void deleteNotice(Long notice_idx) {
         noticeRepository.deleteById(notice_idx);
         
     }
